Add documentation for the syntax
This also sets a basic goalpost we can implement for now.

The currently documented things are basic arithmetic expressions:
  2 + x + 5 * 5
Basic definitions:
  x = 4 + 4
Top-level type annotations:
  x : I64
  x = 5
(no annotations in other places for now)

Lambdas:
  x = \y -> 4

This is sufficiently rich to warrant interesting type checking and
evaluating, so the next steps would be writing a compiler that can
fulfill all of these things.