Add parsing for arithmetic operators
This adds parsing for operators with precedence and associativity. PEG
makes this very easy to do.

One problem is that we don't handle whitespace between operators. To fix
this, the best recourse is probably adding a lexer phase, and then
passing tokens rather than a string to the parser generator.