Create a lexer
We need a lexer to handle whitespace nicely, for now. Eventually the
lexer will become crucial to implementing indentation sensitive syntax.
For now it's more of a convenience than anything else.

The design of the lexer is pretty similar to previous projects of mine,
like poline. The main differences this time are the streamlined
interface, just providing a single `lex` function that returns either a
list of tokens, or a list of errors, as well as a better implementation.

I managed to make the implementation a bit more elegant than what I was
used to doing, by moving out skipping whitespace to a seperate function.
This means that you know that all whitespace has been skipped in the
main `next` function, so you don't need to do `return
Some(Ok(Token::T))` at every point where a token is generated, since
you're not embedded in a while loop. This makes the declaration a lot
more succint, since there's no loop inside the next call itself, just a
match returning either Ok or Err at each branch. We then wrap that match
into Some automatically.

One issue I realized is that handling unary minus is a bit tricky, and
will require making things smart inside of this parser. I forgot that
haskell had these issues about `-` but they are present. For example,
`map (-2)` interprets the `-2` as unary minus applied to 2, and not as
the function `\x -> x - 2`, which can be annoying at times. Haskell also
interprets `f -4` as `f - 4`, which is a bit easier to parse.