Allow unary minus syntactically
This edits the description of the syntax to allow for unary minus:
  f (-x)

This is technically ambiguous, but the PEG rule of always accepting the
first match means that expressions like:
  - x - y
will be parsed in the "least bad" way, that is as
  (-x) - y
The parser will find:
  sub(neg(x), y)
and:
  apply(neg(x), neg(y))

but the structure makes it so that the first is matched before the
second, so we get the right thing.

Need to be careful that the parser is implemented correctly about this
aspect though.