Implement a full parser for the language
A few tweaks were necessary to make things work. Namely, we now require
semicolons between definitions. Take this program as an example
  x = 2
  y = 4
Without a semicolon we read:
  x = 2 y
x is the function `2` applied to the argument `y`, but then we don't
know how to interpret the
  = 4
that remains.

If you require semicolons between definitions this solves the problem.
To get back to the usual syntax, we need to have implicit semicolon
insertion.