Create a more complex structure of types in the typer
This introduces a full tree allowing us represent uncertainty precisely
while type-checking. So we can have the type (? -> I64) -> String, which
represents a type where just a single part of the function is not yet
known. This is necessary to be able to elaborate different parts as we
come onto them.

To do this while minimizing repetition, we divide the type structure
into the base types, and then the known combinators (just ->). We end up
with just the concrete types as base types eventually, but while type
checking we also include ? as a type.