Handle function application in type checking
This one is actually a bit tricky. The idea is that we process both the
function and the argument, and then we make sure that the function type
we've inferred looks like A -> ? where A is the argument type we've
inferred. Then the result of the expression is whatever we've managed to
infer for the final result of the function.