Allow typing of lambda expressions
This is the first use of contexts so far. We introduce a new error for
when we're not able to infer a type for the argument of the lambda
expression from the usage inside the lambda alone. I think that this
might be able to be relaxed a bit, to end up with a partial type.