Add parsing of polymorphic schemes / types
This has to change quite a few bits to get this simple thing working.
But now a type like:

  const : {a, b} => a -> b -> b

*parses*. Of course, we haven't implemented any type-checking related to
this, so it doesn't work, but it's a first step towards that.

The next step would be to modify the type checker to support this. I
think the type-checker needs to be written towards a separate
constraint-gathering -> unification stage, which would be more in line
with Hindey-Milner type-checking, which this is now powerful enough to
be more-or-less. A lot of ad-hoc logic we had is going to be addressed
by this as well.