Allow pretty printing of parsed ast
This makes it so that we can also choose to display the parsed AST
before simplification.

Now we print out a nice sexp based AST, which is much easier to parse.
In the future we might want to also have a debug option, in order to be
able to choose between the pretty and raw ASTs